requirements.txt
dj-database-url==0.5.0
Django==4.1.2
django-on-heroku==1.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
entrypoints==0.3
evdev==1.6.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
jsonschema==3.2.0
keyboard==0.13.5
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.4
pycairo==1.16.2
pycodestyle==2.9.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynput==1.7.6
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.15.5
python-dateutil==2.7.3
python-xlib==0.31
python-dotenv==0.21.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
systemd-python==234
texttable==1.6.2
toml==0.10.2
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
whitenoise==6.2.0
zipp==1.0.0